Add basket and basket details entites.

+++ Phone-Store-Management/DAO/IDao.cs
@@ -8,7 +8,7 @@ namespace Phone_Store_Management.DAO
 {
     public interface IDao<T>
     {
-        T Get(long id);
+        T Get(int id);
 
         List<T> GetAll();
